The unified XPath- en XQueryContainsAssertions seems to be working. Furthermore fixed an UI bug regarding ignoreComments. The UI did not properly reflect the this setting.
